Project Wing/requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.1.0.dev0
bottle==0.12.13
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cntk==2.6
cntk-gpu==2.6
colorama==0.4.0
cycler==0.10.0
decorator==4.3.0
Django==2.1.3
docutils==0.14
et-xmlfile==1.0.1
Flask==1.0.2
gast==0.2.0
geographiclib==1.49
geopy==1.17.0
googlemaps==3.0.2
grpcio==1.16.0
h5py==2.8.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kivy.deps.sdl2==0.1.18
Kivy==1.10.1
Kivy-Garden==0.1.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.1
mccabe==0.6.1
networkx==2.2
numpy==1.15.4
openpyxl==2.5.9
pandas==0.23.4
Pillow==5.3.0
pip==18.1
progressbar2==3.38.0
protobuf==3.6.1
Pygments==2.2.0
pyjade==4.0.0
pylint==2.1.1
pyparsing==2.3.0
pypiwin32==223
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
pythonnet==2.3.0
python-utils==2.3.0
pytz==2018.7
pywin32==224
PyYAML==3.13
requests==2.20.0
scikit-learn==0.20.0
scipy==1.1.0
setuptools==40.5.0
sip==4.19.8
six==1.11.0
tensorboard==1.12.0
tensorflow==1.11.0
termcolor==1.1.0
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wheel==0.32.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
